Updating a number of dependencies Fixes some issues with @IntDefs using long in some cases, also fixes a couple of method signatures to properly accept integers Fixes some nullability issues after upgrading to the latest support library

+++ app/src/main/kotlin/com/kelsos/mbrc/content/library/albums/AlbumRepositoryImpl.kt
@@ -42,7 +42,7 @@ constructor(
 
   override fun cacheIsEmpty(): Single<Boolean> = Single.fromCallable { dao.count() == 0L }
 
-  override fun getAlbumsSorted(@Sorting.Fields order: Long, ascending: Boolean): Single<DataSource.Factory<Int, AlbumEntity>> {
+  override fun getAlbumsSorted(@Sorting.Fields order: Int, ascending: Boolean): Single<DataSource.Factory<Int, AlbumEntity>> {
     val factory = when (order) {
       Sorting.ALBUM -> {
         if (ascending) {
